backend/requirements.txt
fastapi==0.117.1
langchain_openai==0.3.33
langgraph==0.6.7
pypdf==6.1.0
python-dotenv==1.1.1
python_docx==1.2.0
transformers==4.55.0
